refactor: Eliminate compiler warning
The compiler complains during the integration test because the main()
function isn't used. That happens because the launch macro generates the
main function regardless of testing, while in the test executable the
main function is never called.

Having an explicit main allows to disable the code in the test
configuration.